Fix heat gauge using int arithmetic where it should be using floats

+++ src/main/java/xyz/phanta/unclunkify/tile/base/TileProcessing.java
@@ -92,7 +92,7 @@ public abstract class TileProcessing<R extends IRcp<ItemStack, ItemStackInput, I
     }
 
     public float getHeatFraction() {
-        return heat.getInt() / getMaxHeatTicks();
+        return heat.getInt() / (float)getMaxHeatTicks();
     }
 
     public float getWorkFraction() {
